refactor: update PrometheusHandler to use a custom registry
This commit updates the PrometheusHandler function in the handler package to use a custom registry instead of the default one. The custom registry is passed as a parameter to the function. This change allows for more flexibility in managing and registering Prometheus metrics.

The changes were made in the following files:
- handler/handler.go
- server.go